Split wgsl_compile.rs
