+++ crates/darkly/src/brush/nodes/paint_color.rs
@@ -7,7 +7,7 @@ use std::sync::Arc;
 
 use crate::brush::eval::{BrushNodeEvaluator, EvalContext};
 use crate::brush::node::BrushNodeRegistration;
-use crate::brush::wgsl_compile::{CompileWgslCtx, NodeWgsl, UniformField, WgslType};
+use crate::brush::wgsl::{CompileWgslCtx, NodeWgsl, UniformField, WgslType};
 use crate::brush::wire::BrushWireType;
 use crate::brush::wire::ScalarValue;
 use crate::nodegraph::{NodeRegistration, PortDef};
